commiting local changes to setup_appdir.sh

+++ setup_appdir.sh
@@ -1,21 +1,30 @@
 #!/bin/bash
+set -e
 
-wget -O dist/python3.11.AppImage https://github.com/niess/python-appimage/releases/download\
-/python3.11/python3.11.1-cp311-cp311-manylinux_2_24_x86_64.AppImage
+# Get the directory of the script
+DIR="$( cd "$( dirname "${BASH_SOURCE[0]}" )" &> /dev/null && pwd )"
+
+# create dist dir
+[ -d "$DIR/dist" ] || mkdir "$DIR/dist"
+
+wget -O "$DIR/dist/python3.11.AppImage" https://github.com/niess/python-appimage/releases/download/python3.11/python3.11.1-cp311-cp311-manylinux_2_24_x86_64.AppImage
 
 # get appbuildtool
-wget -O dist/appimagetool.AppImage https://github.com/AppImage/AppImageKit/releases/download/continuous/\
-appimagetool-x86_64.AppImage
+wget -O "$DIR/dist/appimagetool.AppImage" https://github.com/AppImage/AppImageKit/releases/download/continuous/appimagetool-x86_64.AppImage
 
+# ensure that the appimage files exist
+[ -f "$DIR/dist/appimagetool.AppImage" ] || exit 1
+[ -f "$DIR/dist/python3.11.AppImage" ] || exit 1
 
 # make app image executable
-chmod +x dist/appimagetool.AppImage
-chmod +x dist/python3.11.AppImage
+sudo chmod +x "$DIR/dist/appimagetool.AppImage"
+sudo chmod +x "$DIR/dist/python3.11.AppImage"
 
 # extract python app image
-./dist/python3.11.AppImage --appimage-extract
+ls -la "$DIR/dist"
+"$DIR/dist/python3.11.AppImage" --appimage-extract
 
-[ -d "AppDir" ] && rm -r AppDir
+[ -d "$DIR/AppDir" ] && rm -r "$DIR/AppDir"
 
 # overwrite files from copy dir
 mv squashfs-root AppDir
@@ -29,6 +38,6 @@ cp dist/CaptainsLog AppDir/usr/bin
 rm AppDir/usr/share/metainfo/python*
 
 # re-package app image
-./dist/appimagetool.AppImage ./AppDir ./dist/CaptainsLog.AppImage 
+dist/appimagetool.AppImage AppDir dist/CaptainsLog.AppImage
 
 # potentially upload?
